fix: when converting int8/uint8 to str, cast to int

+++ include/xtypes/DynamicDataImpl.hpp
@@ -148,10 +148,10 @@ inline std::string ReadableDynamicDataRef::to_string() const
                 ss << "<" << type_name << ">  " << static_cast<int32_t>(node.data().value<wchar_t>());
                 break;
             case TypeKind::INT_8_TYPE:
-                ss << "<" << type_name << ">  " << node.data().value<int8_t>();
+                ss << "<" << type_name << ">  " << static_cast<int>(node.data().value<int8_t>());
                 break;
             case TypeKind::UINT_8_TYPE:
-                ss << "<" << type_name << ">  " << node.data().value<uint8_t>();
+                ss << "<" << type_name << ">  " << static_cast<int>(node.data().value<uint8_t>());
                 break;
             case TypeKind::INT_16_TYPE:
                 ss << "<" << type_name << ">  " << node.data().value<int16_t>();
